Add the actual NED package - just system model :/
Adding all models fails compilation. I just need a small NED right now,
so I've gone ahead and added just the system model and its
dependencies. We should add in the rest later on. I included the YANG
models in the ietf-yangs directory so it's easy to add in later on.